feat(cedar e2e tests): move all nightwatch globals to nightwatch.conf.js. Delete globals.js file
Move all configs from the globals.js file to nightwatch.conf.js. Remove unused nightwatch-vite
plugin.

CDR-2520
