+++ package.json
@@ -61,7 +61,7 @@
     "backstopjs": "^5.3.0",
     "c8": "^7.11.0",
     "chalk": "^4.1.0",
-    "chromedriver": "^99.0.0",
+    "chromedriver": "^101.0.0",
     "concat": "^1.0.3",
     "cross-env": "^7.0.3",
     "cz-conventional-changelog": "^3.3.0",
@@ -99,7 +99,6 @@
     "stylelint-config-recommended-scss": "^4.2.0",
     "stylelint-scss": "^3.19.0",
     "vite": "^2.9.1",
-    "vite-plugin-nightwatch": "^0.2.12",
     "vitest": "^0.8.1",
     "vue-router": "^4.0.10"
   },
