+++ test/e2e/nightwatch.conf.js
@@ -1,15 +1,24 @@
 //require('@babel/register');
 
 // http://nightwatchjs.org/getingstarted#settings-file
+
+const vite = require('./vite.js');
+
+const startViteServer = function() {
+  return vite.start({
+    configFile: './vite.config.js',
+  });
+}
+
+let viteServer;
+
 module.exports = {
-  plugins: ['vite-plugin-nightwatch'],
   src_folders: ['test/e2e/specs'],
   output_folder: 'test/e2e/reports',
   custom_assertions_path: ['test/e2e/custom-assertions'],
   custom_commands_path: [
     './node_modules/nightwatch-axe/src/commands',
   ],
-  globals_path: './globals.js',
 
   test_settings: {
     default: {
@@ -17,6 +26,14 @@ module.exports = {
       globals: {
         devServerURL: `http://localhost:3000/#/`,
         asyncHookTimeout: 180000,
+        async before() {
+          viteServer = await startViteServer(); 
+          const port = viteServer.config.server.port;
+          this.launchUrl = `http://localhost:${port}`;
+        },
+        async after() {
+          await viteServer.close();
+        },
       },
       desiredCapabilities: {
         browserName : 'chrome'
